Now applying best guess for plan type (IFR/VFR) for flight plan formats not supporting this.
Fixed several issues when determining cruise altitude.
Rounding error related procedure altitude violations corrected when using metric altitude.
Fixed issues with wrong flight plan cruise altitude when using metric altitude. Rounding errors when
loading and saving can still appear since all flight plan formats use feet.

+++ src/common/unit.cpp
@@ -1,5 +1,5 @@
 /*****************************************************************************
-* Copyright 2015-2020 Alexander Barthel alex@littlenavmap.org
+* Copyright 2015-2023 Alexander Barthel alex@littlenavmap.org
 *
 * This program is free software: you can redistribute it and/or modify
 * it under the terms of the GNU General Public License as published by
@@ -407,7 +407,7 @@ float Unit::altFeetF(float value)
 
 int Unit::altFeetI(int value)
 {
-  return atools::roundToInt(altMeterF(ageo::feetToMeter(value)));
+  return atools::roundToInt(altMeterF(ageo::feetToMeter(static_cast<float>(value))));
 }
 
 QString Unit::volGallon(float value, bool addUnit)
